requirements.txt
# Core dependencies
langchain==0.0.350
streamlit==1.28.1
python-dotenv==1.0.0

# Hugging Face
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2

# Vector stores
pinecone-client==2.2.4

# Document processing
PyPDF2==3.0.1

# Utilities (LangChain compatible)
numpy==1.26.4
pydantic==2.5.0
